Refactor directory structure; move feature files to a more organized app directory and update import paths for improved maintainability.
