+++ lib/ui/assets/widgets/tree/assets_tree_isolate.dart
@@ -1,8 +1,8 @@
 import 'dart:isolate';
 
-import '../../../../../core/domain/company_asset/company_asset.dart';
-import '../../../../../core/domain/company_location/company_location.dart';
-import '../../../../../core/domain/uid.dart';
+import '../../../../core/domain/company_asset/company_asset.dart';
+import '../../../../core/domain/company_location/company_location.dart';
+import '../../../../core/domain/uid.dart';
 import 'assets_tree_filters.dart';
 import 'assets_tree_state.dart';
