+++ lib/ui/assets/widgets/tree/assets_tree_state.dart
@@ -1,8 +1,8 @@
 import 'package:equatable/equatable.dart';
 
-import '../../../../../../core/domain/uid.dart';
-import '../../../../../core/domain/company_asset/company_asset.dart';
-import '../../../../../core/domain/company_location/company_location.dart';
+import '../../../../../core/domain/uid.dart';
+import '../../../../core/domain/company_asset/company_asset.dart';
+import '../../../../core/domain/company_location/company_location.dart';
 import 'assets_tree_filters.dart';
 import 'assets_tree_node_model.dart';
