+++ lib/ui/assets/widgets/tree/node/asset_node_icon.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 
-import '../../../../../../core/domain/company_asset/company_asset.dart';
-import '../../../../../../core/domain/company_location/company_location.dart';
+import '../../../../../core/domain/company_asset/company_asset.dart';
+import '../../../../../core/domain/company_location/company_location.dart';
 import '../../../../app/widgets/app_icon.dart';
 import '../assets_tree_node_model.dart';
