+++ lib/ui/assets/widgets/tree/node/asset_node_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../../../../core/domain/company_asset/company_asset.dart';
+import '../../../../../core/domain/company_asset/company_asset.dart';
 import '../../../../app/styles/styles.dart';
 import '../assets_tree_node_model.dart';
 import 'asset_node_icon.dart';
