+++ lib/ui/assets/widgets/tree/node/asset_tree_vertical_line.dart
